+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 All notable changes to this project will be documented in this file.
+### [0.10.1] - 2018-09-27
+[fix] get more messages button duplicated
+[fix] hexViewer symbols <0x20 and >0x7f
 ### [0.10.0] - 2018-09-26
 [feature] rebased hex viewer
 [feature] time selector added to channels messages